+++ package.json
@@ -13,20 +13,20 @@
     "lint:fix": "eslint --ext .js,.vue --ignore-path .gitignore . --fix"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.12.5",
+    "@nuxtjs/axios": "^5.13.0",
     "core-js": "^3.8.3",
     "feather-icons": "^4.28.0",
     "nuxt": "^2.14.12",
-    "tauri": "^0.14.0",
+    "tauri": "^0.14.1",
     "vue": "^2.6.12",
     "vue-feather": "^1.1.1"
   },
   "devDependencies": {
-    "@nuxtjs/color-mode": "^2.0.2",
+    "@nuxtjs/color-mode": "^2.0.3",
     "@nuxtjs/eslint-config": "^5.0.0",
     "@nuxtjs/eslint-module": "^3.0.2",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.18.0",
+    "eslint": "^7.19.0",
     "eslint-plugin-nuxt": "^2.0.0",
     "eslint-plugin-vue": "^7.5.0"
   }
