+++ src-tauri/Cargo.toml
@@ -12,11 +12,8 @@ build = "src/build.rs"
 [dependencies]
 zip = "0.5"
 warp = "0.2"
-actix-rt = "1.1"
-actix-web = "2.0"
 serde_json = "1.0"
 env_logger = "0.7"
-actix-files = "0.2"
 webview_official = "0.0"
 serde = { version = "1.0", features = [ "derive" ] }
 tauri = { version = "0.11", features = [ "all-api" ] }
