--- src-tauri/src/config.rs
@@ -1,3 +1,0 @@
-pub const WS_SERVER_PORT: u16 = 41431;
-pub const GAME_SERVER_PORT: u16 = 41432;
-pub const UDP_SERVER_PORT: u16 = 41433;
