+++ src-tauri/src/constants.rs
@@ -1,0 +1,7 @@
+pub const UDP_PORT: u16 = 41431;
+pub const SERVER_PORT: u16 = 41430;
+
+pub const GAMES_DIR: &str = ".sinix/games";
+pub const GAMES_DATA_DIR: &str = ".sinix/data";
+pub const TEMP_DIR: &str = ".sinix/games/.tmp_game";
+pub const SINIX_CONFIG_FILENAME: &str = "sinix.manifest.json";
