+++ src-tauri/src/plugins/mod.rs
@@ -1,5 +1,7 @@
 mod game_webview;
 mod udp_server;
+mod root;
 
+pub use root::SinixRoot;
 pub use game_webview::GameWebview;
 pub use udp_server::UdpSocketServer;
