+++ src-tauri/src/plugins/udp_server.rs
@@ -7,7 +7,7 @@ use std::sync::mpsc::Sender;
 use serde::Serialize;
 use tauri::{plugin::Plugin, Webview};
 
-use crate::config;
+use crate::constants as sinix_constants;
 
 #[derive(Serialize)]
 pub struct Reply {
@@ -67,7 +67,7 @@ pub fn handle(rx: Receiver<String>){
 }
 
 pub fn serve(tx: Sender<String>){
-  let addr = format!("0.0.0.0:{}", config::UDP_SERVER_PORT);
+  let addr = format!("0.0.0.0:{}", sinix_constants::UDP_PORT);
   let socket = match UdpSocket::bind(addr) {
     Ok(s) => s,
     Err(e) => panic!("couldn't bind socket: {}", e)
