+++ src-tauri/src/routes/channel.rs
@@ -1,13 +1,9 @@
-use env_logger;
-use crate::config;
-
 use std::collections::HashMap;
 use std::convert::Infallible;
 use std::sync::Arc;
 
 use tokio::sync::{mpsc, RwLock};
-use tokio::runtime;
-use warp::{ws::Message, Filter, Rejection};
+use warp::{ws::Message, Filter, Rejection, Reply};
 
 mod handler;
 mod ws;
@@ -21,19 +17,8 @@ pub struct Client {
   pub sender: Option<mpsc::UnboundedSender<std::result::Result<Message, warp::Error>>>,
 }
 
-pub fn init() {
-  tauri::spawn(move || {
-    println!("websocket: {:?}", std::thread::current().id());
-    runtime::Runtime::new().unwrap().block_on(async {
-      serve().await;
-    });
-  });
-}
-
-pub async fn serve() {
-  env_logger::init();
+pub fn route() -> impl Filter<Extract = (impl Reply,), Error = warp::Rejection> + Clone {
   let clients: Clients = Arc::new(RwLock::new(HashMap::new()));
-  let health_route = warp::path!("health").and_then(handler::health_handler);
 
   let register = warp::path("register");
   let register_routes = register
@@ -60,19 +45,12 @@ pub async fn serve() {
     .and(with_clients(clients.clone()))
     .and_then(handler::ws_handler);
 
-  let routes = health_route
-    .or(ws_route)
-    .or(register_routes)
-    .or(publish)
-    .with(
-      warp::cors()
-        .allow_any_origin()
-        .allow_methods(vec!["GET", "POST", "DELETE", "OPTION"]),
-    );
-
-  warp::serve(routes)
-    .run(([0, 0, 0, 0], config::WS_SERVER_PORT))
-    .await;
+  warp::path("channel")
+  .and(
+    register_routes
+      .or(publish)
+      .or(ws_route)
+  )
 }
 
 fn with_clients(clients: Clients) -> impl Filter<Extract = (Clients,), Error = Infallible> + Clone {
