--- src-tauri/src/services/game/config.rs
@@ -1,3 +1,0 @@
-pub const GAMES_DIR: &str = ".sinix/games";
-pub const TEMP_DIR: &str = ".sinix/games/.tmp_game";
-pub const SINIX_CONFIG_FILENAME: &str = "sinix.manifest.json";
